Get weather data (#8)
* feat: Get weather data (OpenWeather)

We can get today's weather data in Tokyo.
This feature uses OpenWeather API (One Call API 3.0) to get the weather
data. If you want to know it, you should create your account of
OpenWeather and get your API.

https://openweathermap.org/api/one-call-3

This separate subscription includes 1,000 calls/day for free and allows
you to pay only for the number of API calls made to this product.
Once you subscribe to One call API 3.0, 2000 API calls per day to this
product are set up by default. If you want to change this limit, please
go to the "Billing plans" tab in your Personal account to update
standard settings.

* build: Update build environment (Docker, CMake)

* feat: Formatting stdout info

Formatting the output of the weather data to use the `fmt` library.

* feat: Change temperature unit (fahrenheit->celsius)

The default temperature unit is `Fahrenheit`.
By adding `&units=metric` to the end of the URL(API call),
the temperature can be responded in `Celsius` units.

* build: Setup build environment for `fmt`

* fix: Remove build artifacts forcibly (sudo)

* feat: Helper script for Docker

* refactor: Change the path of the script to call

* refactor: Change the stdout message

* refactor: Remove unnecessary steps

The build process was divided as follows:

- Build with CMake
- Build and deploy docs with MkDocs

* feat: Build with CMake in Docker container

* test: Build workflow

* refactor: Change the message

* fix: Fixed docker option

Remove `-i` option.